+++ docs/src/release_notes.md
@@ -1,10 +1,13 @@
 # Release Notes
 These release notes adhere to the [keep a changelog](https://keepachangelog.com/en/1.0.0/) format. Below is a list of changes since CausalELM was first released.
 
-## Version [0.8.0](https://github.com/dscolby/CausalELM.jl/releases/tag/v0.8.0) - 2024-10-28
+## Version [0.8.0](https://github.com/dscolby/CausalELM.jl/releases/tag/v0.8.0) - 2024-10-31
+### Added
+*   Implemented randomization inference-based confidence intervals [#78](https://github.com/dscolby/CausalELM.jl/issues/78)
 ### Fixed
 *   Removed unnecessary include and using statements
-*   Slightly sped up the randomizatioin inference implementation and clarified it in the docs [#77](https://github.com/dscolby/CausalELM.jl/issues/77)
+*   Slightly sped up the randomization inference implementation and clarified it in the docs [#77](https://github.com/dscolby/CausalELM.jl/issues/77)
+*   Fixed the randomization inference index selection procedure for interrupted time series estimators
 
 ## Version [v0.7.0](https://github.com/dscolby/CausalELM.jl/releases/tag/v0.7.0) - 2024-06-22
 ### Added